requirements.txt
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
streamlit==1.36.0
streamlit_authenticator==0.3.2
